+++ demos/unpaired_mr_brain/demo_predict.py
@@ -34,17 +34,17 @@ print(
     "deepreg_predict --gpu '' "
     f"--config_path demos/{name}/{name}.yaml "
     f"--ckpt_path demos/{name}/dataset/pretrained/learn2reg_t4_unpaired_weakly_train_logs2/save/weights-epoch200.ckpt "
-    f"--log_root demos/{name} "
-    "--log_dir logs_predict "
+    f"--log_dir demos/{name} "
+    "--exp_name logs_predict "
     "--save_png --mode test\n"
     "=========================================================\n"
     "\n\n\n\n\n"
 )
 
-log_root = f"demos/{name}"
-log_dir = "logs_predict/" + datetime.now().strftime("%Y%m%d-%H%M%S")
-ckpt_path = f"{log_root}/dataset/pretrained/learn2reg_t4_unpaired_weakly_train_logs2/save/weights-epoch200.ckpt"
-config_path = [f"{log_root}/{name}.yaml"]
+log_dir = f"demos/{name}"
+exp_name = "logs_predict/" + datetime.now().strftime("%Y%m%d-%H%M%S")
+ckpt_path = f"{log_dir}/dataset/pretrained/learn2reg_t4_unpaired_weakly_train_logs2/save/weights-epoch200.ckpt"
+config_path = [f"{log_dir}/{name}.yaml"]
 if args.test:
     config_path.append("config/test/demo_unpaired_grouped.yaml")
 
@@ -54,8 +54,7 @@ predict(
     ckpt_path=ckpt_path,
     mode="test",
     batch_size=1,
-    log_root=log_root,
     log_dir=log_dir,
-    sample_label="all",
+    exp_name=exp_name,
     config_path=config_path,
 )