+++ test/unit/test_predict.py
@@ -10,8 +10,6 @@ import shutil
 
 from deepreg.predict import build_config, build_pair_output_path
 
-log_root = "logs"
-
 
 def test_build_pair_output_path():
     """
@@ -42,22 +40,23 @@ def test_build_pair_output_path():
 
 def test_build_config():
     """
-    Test build_config and check log_dir setting and checkpoint path verification
+    Test build_config and check exp_name setting and checkpoint path verification
     """
     config_path = "config/unpaired_labeled_ddf.yaml"
-    log_dir = "test_build_config"
+    exp_name = "test_build_config"
+    log_dir = "logs"
 
     # TODO checkpoint path empty
 
     # checkpoint path ends with ckpt
     got_config, got_log_dir, _ = build_config(
         config_path=config_path,
-        log_root=log_root,
         log_dir=log_dir,
+        exp_name=exp_name,
         ckpt_path="example.ckpt",
     )
     assert isinstance(got_config, dict)
-    assert got_log_dir == os.path.join("logs", log_dir)
+    assert got_log_dir == os.path.join(log_dir, exp_name)
 
 
 def test_predict_on_dataset():